+++ src/models/Roles.ts
@@ -1,6 +1,6 @@
-import { RolesApiListRolesRequest } from "sailpoint-api-client";
+import { RolesV2025ApiListRolesRequest } from "sailpoint-api-client";
 
-export const DEFAULT_ROLES_QUERY_PARAMS: RolesApiListRolesRequest = {
+export const DEFAULT_ROLES_QUERY_PARAMS: RolesV2025ApiListRolesRequest = {
     count: false,
     limit: 250,
     offset: 0,